Separate modules for lexer and parser
